% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{ACCIDENT}
\alias{ACCIDENT}
\title{Monthly numbers of road traffic accidents with personal injury in BRD}
\format{ACCIDENT is a univariate time series of length 528, start  January 1974, frequency = 12   
\describe{
\item{ACCIDENT}{Monthly  numbers of  road traffic accidents with personal injury}
}}
\source{
< https://www-genesis.destatis.de/genesis//online?operation=table&code=46241-0002&levelindex=0&levelid=1583749114977>
}
\usage{
ACCIDENT
}
\description{
Monthly numbers of road traffic accidents with personal injury in BRD
}
\examples{
data(ACCIDENT)
## maybe  tsp(ACCIDENT) ; plot(ACCIDENT) 
}
\keyword{datasets}
