% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarts.r
\name{acfmat}
\alias{acfmat}
\title{\code{acfmat} computes a sequence of autocorrelation matrices for a multivariate time series}
\usage{
acfmat(y, lag.max)
}
\arguments{
\item{y}{multivariate time series}

\item{lag.max}{maximum number of lag}
}
\value{
out   list with components: 
\item{M}{ array with autocovariance matrices }
\item{M1}{array with indicators if autocovariances are significantly greater (+),    
                 lower (-) than the critical value or insignificant (.) at 95 percent level        }
}
\description{
\code{acfmat} computes a sequence of autocorrelation matrices for a multivariate time series
}
\examples{
data(ICECREAM)
out <- acfmat(ICECREAM,7) 
}
