% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smaniplot.r
\name{acfpacf}
\alias{acfpacf}
\title{\code{acfpacf} produces a plot of the acf and the pacf of a time series}
\usage{
acfpacf(x, lag, HV = "H")
}
\arguments{
\item{x}{the series, a vector or a time series}

\item{lag}{scalar, maximal lag to be plotted}

\item{HV}{character, controls division of graphic window: "H" horizontal, "V" vertical, default is "H"}
}
\description{
\code{acfpacf} produces a plot of the acf and the pacf of a time series
}
\examples{
\donttest{
  data(LYNX)
  acfpacf(log(LYNX),15,HV="H") }
}
