% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequdom.r
\name{armathspec}
\alias{armathspec}
\title{\code{armathspec} determines the theoretical spectrum of an arma process}
\usage{
armathspec(a, b, nf, s = 1, pl = FALSE)
}
\arguments{
\item{a}{ar-coefficients}

\item{b}{ma-coefficients}

\item{nf}{scalar, the number of equally spaced frequencies}

\item{s}{variance of error process}

\item{pl}{logical,  if TRUE, the spectrum is plotted, FALSE for no plot}
}
\value{
out   (nf+1,2) matrix, the frequencies and the spectrum
}
\description{
\code{armathspec} determines the theoretical spectrum of an arma process
}
\examples{
out <-armathspec(c(0.3,-0.5),c(-0.8,0.7),50,s=1,pl=FALSE) 

}
