% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.r
\name{aspectratio}
\alias{aspectratio}
\title{\code{aspectratio}  determines the aspect ratio to plot a time series}
\usage{
aspectratio(y)
}
\arguments{
\item{y}{time series}
}
\value{
a      scalar, the aspect ratio
}
\description{
\code{aspectratio}  determines the aspect ratio to plot a time series
}
\examples{
data(GDP)
a <- aspectratio(GDP) 
}
