% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{BEER}
\alias{BEER}
\title{Monthly beer production in Australia: megalitres. Includes ale and stout. Does not include beverages with alcohol percentage less than 1.15.}
\format{
BEER is a univariate time series of length 476, start January 1956, end Aug 1995, frequency = 12
\describe{
\item{BEER}{Monthly production of beer in Australia}
}
}
\source{
R package tsdl  <https://github.com/FinYang/tsdl>
}
\usage{
BEER
}
\description{
Monthly beer production in Australia: megalitres. Includes ale and stout. Does not include beverages with alcohol percentage less than 1.15.
}
\examples{
data(BEER)
## maybe  tsp(BEER) ; plot(BEER) 
}
\keyword{datasets}
