% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smaniplot.r
\name{BoxCox}
\alias{BoxCox}
\title{\code{BoxCox} determines the power of a Box-Cox transformation to stabilize the variance of a time series}
\usage{
BoxCox(y, seg, Plot = FALSE)
}
\arguments{
\item{y}{the series, a vector or a time series}

\item{seg}{scalar, number of segments}

\item{Plot}{logical, should a plot be produced?}
}
\value{
l    scalar, the power of the Box-Cox transformation
}
\description{
\code{BoxCox} determines the power of a Box-Cox transformation to stabilize the variance of a time series
}
\examples{
data(INORDER)
lambda <-BoxCox(INORDER,6,Plot=FALSE)

}
