% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{COFFEE}
\alias{COFFEE}
\title{U.S. annual coffee consumption}
\format{
COFFEE is a univariate time series of length 61; start 1910, frequency = 1 
\describe{
\item{COFFEE}{annual coffee-consumption USA, logarithmic transformed}
}
}
\source{
R package tsdl <https://github.com/FinYang/tsdl>
}
\usage{
COFFEE
}
\description{
U.S. annual coffee consumption
}
\examples{
data(COFFEE)
## maybe  tsp(COFFEE) ; plot(COFFEE) 
}
\keyword{datasets}
