% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{DAX}
\alias{DAX}
\title{Market value of DAX}
\format{
DAX is a multivariate time series of length 12180 and 4 variables
\describe{
\item{DAY}{Day of the week} 
\item{MONTH}{Month} 
\item{Year}{Year} 
\item{DAX30}{Market value} 
}
}
\usage{
DAX
}
\description{
Market value of DAX
}
\examples{
data(DAX)
## maybe  tsp(DAX) ; plot(DAX) 
}
\keyword{datasets}
