% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{ENGINES}
\alias{ENGINES}
\title{ENGINES is an alias for MACHINES}
\format{
ENGINES is a univariate time series of length 188, start January 1972 frequency = 12
\describe{
\item{ENGINES}{Incoming orders for engines}
}
}
\usage{
ENGINES
}
\description{
ENGINES is an alias for MACHINES
}
\examples{
data(ENGINES)
## maybe  tsp(ENGINES) ; plot(ENGINES)  
}
\keyword{datasets}
