% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{GDP}
\alias{GDP}
\title{Germany's gross domestic product adjusted for price changes}
\format{GDP is a univariate time series of length 159, start January 1970, frequency = 4
\describe{
\item{GDP}{Gross domestic product  adjusted for price changes}
}}
\source{
<https://www-genesis.destatis.de/genesis//online?operation=table&code=81000-0002&levelindex=0&levelid=1583750132341>
}
\usage{
GDP
}
\description{
Germany's gross domestic product adjusted for price changes
}
\examples{
data(GDP)
## maybe  tsp(GDP) ; plot(GDP) 
}
\keyword{datasets}
