% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{HEARTBEAT}
\alias{HEARTBEAT}
\title{Cardiac frequency of a patient}
\format{HEARTBEAT is a univariate time series of length 30:
\describe{
\item{HEARTBEAT}{cardiac frequency of a patient}
}}
\usage{
HEARTBEAT
}
\description{
Cardiac frequency of a patient
}
\examples{
data(HEARTBEAT)
## maybe  tsp(HEARTBEAT) ; plot(HEARTBEAT) 
}
\keyword{datasets}
