% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{HSV}
\alias{HSV}
\title{HSV's position in the first German soccer league}
\format{HSV is a univariate time series of length 34:
\describe{
\item{HSV}{HSV's position in the first German soccer league}
}}
\source{
<https://www.transfermarkt.de/hamburger-sv/platzierungen/verein/41>
}
\usage{
HSV
}
\description{
HSV's position in the first German soccer league
}
\examples{
data(HSV)
## maybe  tsp(HSV) ; plot(HSV) 
}
\keyword{datasets}
