% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{ICECREAM}
\alias{ICECREAM}
\title{Temperature and consumption of ice cream}
\format{
ICECREAM is a bivariate time series of length 160:  
\describe{ 
\item{ICE}{consumption of ice cream} 
\item{TEMP}{Temperature in Fahrenheit degrees}  
}
}
\source{
Hand, D. J., et al. (1994, ISBN: 9780412399206)  "A Handbook of Small Data Sets"
}
\usage{
ICECREAM
}
\description{
Temperature and consumption of ice cream
}
\examples{
data(ICECREAM)
## maybe  tsp(ICECREAM) ; plot(ICECREAM) 
}
\keyword{datasets}
