% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{INORDER}
\alias{INORDER}
\title{Income orders of a company}
\format{INORDER is a univariate time series of length 237, start January 1968, frequency =12 
\describe{
\item{INORDER}{Income orders of a company}
}}
\usage{
INORDER
}
\description{
Income orders of a company
}
\examples{
data(INORDER)
## maybe  tsp(INORDER) ; plot(INORDER) 
}
\keyword{datasets}
