% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialfun.r
\name{interpol}
\alias{interpol}
\title{\code{interpol} help function for missls}
\usage{
interpol(rho, xcent)
}
\arguments{
\item{rho}{autocorrelation function}

\item{xcent}{centered time series}
}
\value{
z  new version of xcent
}
\description{
\code{interpol} help function for missls
}
