% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hac.r
\name{kweightsHAC}
\alias{kweightsHAC}
\title{\code{kweightsHAC}  help function for HAC}
\usage{
kweightsHAC(
  kernel = c("Truncated", "Bartlett", "Parzen", "Tukey-Hanning", "Quadratic Spectral"),
  dimN,
  bw
)
}
\arguments{
\item{kernel}{kernel function, choose between "Truncated", "Bartlett", "Parzen", "Tukey-Hanning", "Quadratic Spectral".}

\item{dimN}{number of observations}

\item{bw}{bandwidth parameter}
}
\value{
ww weights
}
\description{
\code{kweightsHAC}  help function for HAC
}
