% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{L921}
\alias{L921}
\title{Subsoil water level and precipitation at pilot well L921}
\format{L921 is a trivariate time series of length 335:
\describe{
\item{T}{Day}
\item{Y}{Water level}
\item{Z}{Supplemented water level}
}}
\usage{
L921
}
\description{
Subsoil water level and precipitation at pilot well L921
}
\examples{
data(L921)
## maybe  tsp(L921) ; plot(L921) 
}
\keyword{datasets}
