% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequdom.r
\name{lagwinba}
\alias{lagwinba}
\title{\code{lagwinba} Bartlett's Lag-window for indirect spectrum estimation}
\usage{
lagwinba(NL)
}
\arguments{
\item{NL}{number of lags used for estimation}
}
\value{
win vector, one-sided weights
}
\description{
\code{lagwinba} Bartlett's Lag-window for indirect spectrum estimation
}
\examples{
win <-lagwinba(5) 
}
