% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequdom.r
\name{lagwinpa}
\alias{lagwinpa}
\title{\code{lagwinpa} Parzen's Lag-window for indirect spectrum estimation}
\usage{
lagwinpa(NL)
}
\arguments{
\item{NL}{number of lags used for estimation}
}
\value{
win vector, one-sided weights
}
\description{
\code{lagwinpa} Parzen's Lag-window for indirect spectrum estimation
}
\examples{
win <- lagwinpa(5)  
}
