% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialfun.r
\name{ldrec}
\alias{ldrec}
\title{\code{ldrec} does Levinson-Durbin recursion for determing all coefficients a(i,j)}
\usage{
ldrec(a)
}
\arguments{
\item{a}{(p+1,1)-vector of acf of a time series: acov(0),...,acov(p)
or  1,acor(1),..,acor(p)}
}
\value{
mat (p,p+2)-matrix, coefficients in lower triangular,
 pacf in colum p+2 and Q(p) in colum p+1
}
\description{
\code{ldrec} does Levinson-Durbin recursion for determing all coefficients a(i,j)
}
\examples{
data(HEARTBEAT)
a <- acf(HEARTBEAT,5,plot=FALSE)
mat <- ldrec(a$acf)
}
