% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{LITH}
\alias{LITH}
\title{Daily subsoil water level and precipitation at pilot well Lith}
\format{
LITH is a bivariate time series of length 1347:
\describe{
\item{N}{precipitation amount}    
\item{G}{water level} 
}
}
\usage{
LITH
}
\description{
Daily subsoil water level and precipitation at pilot well Lith
}
\examples{
data(LITH)
## maybe  tsp(LITH) ; plot(LITH) 
}
\keyword{datasets}
