% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{LUHORMONE}
\alias{LUHORMONE}
\title{Level of Luteinzing hormone of a cow}
\format{LUHORMONE is a bivariate time series of length 29:
\describe{
\item{T}{Time in minutes}
\item{X}{Level of the Luteinzing-hormone}
}}
\usage{
LUHORMONE
}
\description{
Level of Luteinzing hormone of a cow
}
\keyword{datasets}
