% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{LYNXHARE}
\alias{LYNXHARE}
\title{Size of populations of lynxes and snow hares}
\format{
LYNXHARE is a simulated bivariate  time series from a VAR[1]-model  of length 100:  
\describe{ 
\item{X}{Number of lynxes} 
\item{Y}{Number of snow hares}  
}
}
\usage{
LYNXHARE
}
\description{
Size of populations of lynxes and snow hares
}
\examples{
data(LYNXHARE)
}
\keyword{datasets}
