% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{MACHINES}
\alias{MACHINES}
\title{Number of incoming orders for machines}
\format{
MACHINES is a univariate time series of length 188, start January 1972 frequency = 12
\describe{
\item{MACHINES}{Incoming orders for machines}
}
}
\usage{
MACHINES
}
\description{
Number of incoming orders for machines
}
\examples{
data(MACHINES)
## maybe  tsp(MACHINES) ; plot(MACHINES)  
}
\keyword{datasets}
