% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{MAUNALOA}
\alias{MAUNALOA}
\title{Atmospheric CO2 concentrations (ppmv) derived from in situ air samples collected at Mauna Loa Observatory, Hawaii}
\format{
MAUNALOA is a univariate time series of length 735; start March 1958, frequency = 12   
\describe{
\item{MAUNALOA}{CO2-concentration at Mauna Loa}
}
}
\source{
Keeling, C. D. ,  Piper, S. C.,  Bacastow,  R. B., Wahlen, M. ,  Whorf, T. P., Heimann,  M.,  and Meijer, H. A.  (2001)  <https://library.ucsd.edu/dc/object/bb3859642r>
}
\usage{
MAUNALOA
}
\description{
Atmospheric CO2 concentrations (ppmv) derived from in situ air samples collected at Mauna Loa Observatory, Hawaii
}
\examples{
data(MAUNALOA)
## maybe  tsp(MAUNALOA) ; plot(MAUNALOA) 
}
\keyword{datasets}
