% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{MDAX}
\alias{MDAX}
\title{Stock market price of MDAX}
\format{MDAX is a multivariate time series of length 6181 and 4 variables
\describe{
\item{DAY}{Day of the week} 
\item{MONTH}{Month} 
\item{YEAR}{Year} 
\item{MDAX}{Opening stock market price} 
}}
\source{
<https://www.onvista.de/index/MDAX-Index-323547>
}
\usage{
MDAX
}
\description{
Stock market price of MDAX
}
\examples{
data(MDAX)
## maybe  tsp(MDAX) ; plot(MDAX[,3]) 
}
\keyword{datasets}
