% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialfun.r
\name{missar}
\alias{missar}
\title{\code{missar} Substitution of missing values in a time series by 
   conditional  exspectations of AR(p) models}
\usage{
missar(x, p, iterout = 0)
}
\arguments{
\item{x}{vector, the time series}

\item{p}{integer, the maximal order of ar polynom   0 <= p < 18, 
if p=0  linear interpolation without iteration}

\item{iterout}{if = 1, iteration history is printed}
}
\value{
out list with   elements
\item{a}{(p,p)-matrix, estimated ar coefficients for ar-models   }
\item{y}{(n,1)-vector, completed time series }
\item{iterhist}{matrix, NULL or the iteration history}
}
\description{
\code{missar} Substitution of missing values in a time series by 
   conditional  exspectations of AR(p) models
}
\examples{
data(HEARTBEAT)
x <- HEARTBEAT
x[c(20,21)] <- NA
out <- missar(x,2)
}
