% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp.r
\name{moveav}
\alias{moveav}
\title{\code{moveav} smoothes a time series by moving averages}
\usage{
moveav(y, q)
}
\arguments{
\item{y}{the series, a vector or a time series}

\item{q}{scalar,  span of moving average}
}
\value{
g vector,  smooth component
}
\description{
\code{moveav} smoothes a time series by moving averages
}
\examples{
data(GDP)
g <- moveav(GDP,12) 
\donttest{ plot(GDP) ; lines(g,col="red") } 
}
