% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{MUSKRAT}
\alias{MUSKRAT}
\title{Annual trade of muskrat pelts}
\format{
MUSKRAT is a univariate time series of length 62; start 1848, frequency = 1 
\describe{
\item{MUSKRAT}{annual trade of muskrat pelts}
}
}
\source{
<https://archive.uea.ac.uk/~gj/book/data/mink.dat>
}
\usage{
MUSKRAT
}
\description{
Annual trade of muskrat pelts
}
\examples{
data(MUSKRAT)
## maybe  tsp(MUSKRAT) ; plot(MUSKRAT) 
}
\keyword{datasets}
