% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{NIKKEI}
\alias{NIKKEI}
\title{Daily values of the Japanese stock market index Nikkei 225 between 02.02.2000 and 20.10.2020}
\format{
NIKKEI is a univariate time series of length 5057
\describe{
\item{NIKKEI}{Daily values of Nikkei}
}
}
\source{
Heber, G., Lunde, A., Shephard, N. and Sheppard, K. (2009) "Oxford-Man Institute's realized library, version 0.3", 
Oxford-Man Institute, University of Oxford, Oxford    <https://realized.oxford-man.ox.ac.uk/data>
}
\usage{
NIKKEI
}
\description{
Daily values of the Japanese stock market index Nikkei 225 between 02.02.2000 and 20.10.2020
}
\examples{
data(NIKKEI)
## maybe plot(NIKKEI) 
}
\keyword{datasets}
