% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{PAPER}
\alias{PAPER}
\title{Two measurements at a paper machine}
\format{
PAPER is a bivariate time series of length 160  
\describe{ 
\item{H}{High} 
\item{W}{Weight}  
}
}
\source{
Janacek, G. J.  & Swift, L. (1993, ISBN: 978-0139184598) "Time Series: Forecasting, Simulation, Applications"
}
\usage{
PAPER
}
\description{
Two measurements at a paper machine
}
\examples{
data(PAPER)
## maybe  tsp(PAPER) ; plot(PAPER) 
}
\keyword{datasets}
