% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequdom.r
\name{periodotest}
\alias{periodotest}
\title{\code{periodotest} computes the p-value of the test for a hidden periodicity}
\usage{
periodotest(y)
}
\arguments{
\item{y}{vector, the time series}
}
\value{
pval the p-value of the test
}
\description{
\code{periodotest} computes the p-value of the test for a hidden periodicity
}
\examples{
  
data(PIGPRICE)
y <- PIGPRICE
out <- stl(y,s.window=6)  
e <- out$time.series[,3]
out <- periodotest(e)
}
