% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequdom.r
\name{perwinba}
\alias{perwinba}
\title{\code{perwinba} Bartlett-Priestley  window for direct spectral estimation}
\usage{
perwinba(e, n)
}
\arguments{
\item{e}{equal bandwidth  (at most n frequencies are used for averaging)}

\item{n}{length of time series}
}
\value{
w   weights (symmetric)
}
\description{
\code{perwinba} Bartlett-Priestley  window for direct spectral estimation
}
\examples{
data(WHORMONE)
w <- perwinba(0.1,length(WHORMONE)) 

}
