% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequdom.r
\name{perwinda}
\alias{perwinda}
\title{\code{perwinda} Daniell  window for direct spectral estimation}
\usage{
perwinda(e, n)
}
\arguments{
\item{e}{equal bandwidth  (at most n frequencies are used for averaging)}

\item{n}{length of time series}
}
\value{
w   weights (symmetric)
}
\description{
\code{perwinda} Daniell  window for direct spectral estimation
}
\examples{
data(WHORMONE)
w <- perwinda(0.1,length(WHORMONE)) 

}
