% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequdom.r
\name{perwinpa}
\alias{perwinpa}
\title{\code{perwinpa} Parzen's window for direct spectral estimation}
\usage{
perwinpa(e, n)
}
\arguments{
\item{e}{equal bandwidth  (at most n frequencies are used for averaging)}

\item{n}{length of time series}
}
\value{
w   weights (symmetric)
}
\description{
\code{perwinpa} Parzen's window for direct spectral estimation
}
\examples{
data(WHORMONE)
w <- perwinpa(0.1,length(WHORMONE)) 

}
