% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialfun.r
\name{pestep}
\alias{pestep}
\title{\code{pestep} help function for missar}
\usage{
pestep(f, xt)
}
\arguments{
\item{f}{IACF, inverse ACF}

\item{xt}{segment of the time series}
}
\value{
xt new version of xt
}
\description{
\code{pestep} help function for missar
}
