% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{PIGPRICE}
\alias{PIGPRICE}
\title{Monthly prices for pigs}
\format{
PIGPRICE is a univariate time series of length 240; start January 1894, frequency =12
\describe{
\item{PIGPRICE}{Monthly prices for pigs}
}
}
\source{
Hanau, A. (1928)  "Die Prognose der Schweinepreise"
}
\usage{
PIGPRICE
}
\description{
Monthly prices for pigs
}
\examples{
data(PIGPRICE)
## maybe  tsp(PIGPRICE) ; plot(PIGPRICE) 
}
\keyword{datasets}
