% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.r
\name{polymake}
\alias{polymake}
\title{\code{polymake} generates the coefficients of an AR process given the zeros of the
characteristic polynomial. The norm of the roots must be greater than  one  for stationary processes.}
\usage{
polymake(r)
}
\arguments{
\item{r}{vector, the zeros of the characteristic polynomial}
}
\value{
C      coefficients (a[1],a[2],...,a[p]) of the polynomial  1 - a[1]z -a[2]z^2 -...- a[p]z^p
}
\description{
\code{polymake} generates the coefficients of an AR process given the zeros of the
characteristic polynomial. The norm of the roots must be greater than  one  for stationary processes.
}
\examples{
C <- polymake(c(2,-1.5,3))

}
