% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{PPDEMAND}
\alias{PPDEMAND}
\title{Peak power demand in Berlin}
\format{
PPDEMAND is a univariate time series of length 37; start 1955, frequency = 1
\describe{
\item{PPDEMAND}{annual peak power demand in Berlin, Megawatt}
}
}
\source{
Fiedler, H. (1979)  "Verschiedene Verfahren zur Prognose des des Stromspitzenbedarfs in Berlin (West)"
}
\usage{
PPDEMAND
}
\description{
Peak power demand in Berlin
}
\examples{
data(PPDEMAND)
## maybe  tsp(PPDEMAND) ; plot(PPDEMAND) 
}
\keyword{datasets}
