% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{PRODINDEX}
\alias{PRODINDEX}
\title{Production index of manufacturing industries}
\format{PRODINDEX is a univariate time series of length 119:
\describe{
  \item{PRODINDEX}{Production index of manufacturing industries}
}}
\source{
Statistisches Bundesamt (2009)  <https://www-genesis.destatis.de/genesis/online>
}
\usage{
PRODINDEX
}
\description{
Production index of manufacturing industries
}
\examples{
data(PRODINDEX)
## maybe  tsp(PRODINDEX) ; plot(PRODINDEX) 
}
\keyword{datasets}
