% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.r
\name{psifair}
\alias{psifair}
\title{\code{psifair} is a psi-function for robust estimation}
\usage{
psifair(u)
}
\arguments{
\item{u}{vector}
}
\value{
out  transformed vector
}
\description{
\code{psifair} is a psi-function for robust estimation
}
\examples{
 
out <- psifair(c(3.3,-0.7,2.1,1.8))

}
