% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.r
\name{psihuber}
\alias{psihuber}
\title{\code{psihuber} is a psi-function for robust estimation}
\usage{
psihuber(u)
}
\arguments{
\item{u}{vector}
}
\value{
out transformed vector
}
\description{
\code{psihuber} is a psi-function for robust estimation
}
\examples{
out <- psihuber(c(3.3,-0.7,2.1,1.8)) 

}
