% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{RAINFALL}
\alias{RAINFALL}
\title{Annual amount of rainfall in Los Angeles}
\format{RAINFALL is a univariate time series of length 119; start 1878, frequency = 1
\describe{
\item{RAINFALL}{Amount of  rainfall in Los Angeles}
}}
\source{
LA Times (January 28. 1997)
}
\usage{
RAINFALL
}
\description{
Annual amount of rainfall in Los Angeles
}
\examples{
data(RAINFALL)
## maybe  tsp(RAINFALL) ; plot(RAINFALL) 
}
\keyword{datasets}
