% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{REDWINE}
\alias{REDWINE}
\title{Monthly sales of Australian red wine (1000 l)}
\format{REDWINE is a univariate time series of length 187; start January 1980, frequency =12
\describe{
\item{REDWINE}{Monthly sales of Australian red wine }
}}
\source{
R package tsdl <https://github.com/FinYang/tsdl>
}
\usage{
REDWINE
}
\description{
Monthly sales of Australian red wine (1000 l)
}
\examples{
data(REDWINE)
## maybe  tsp(REDWINE) ; plot(REDWINE) 
}
\keyword{datasets}
