% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp.r
\name{robsplinedecomp}
\alias{robsplinedecomp}
\title{\code{robsplinedecomp} decomposes a vector into trend, season and irregular  component
by robustified spline approach; a time series attribute is lost}
\usage{
robsplinedecomp(y, d, alpha, beta, Plot = FALSE)
}
\arguments{
\item{y}{the series, a vector or a time series}

\item{d}{seasonal period}

\item{alpha}{smoothing parameter for trend component   (the larger alpha is, the smoother will the smooth component g be)}

\item{beta}{smoothing parameter for  seasonal  component}

\item{Plot}{logical, should a plot be produced?}
}
\value{
out   list with the elements trend, season, residual
}
\description{
\code{robsplinedecomp} decomposes a vector into trend, season and irregular  component
by robustified spline approach; a time series attribute is lost
}
\examples{
data(GDP) 
out  <- robsplinedecomp(GDP,4,2,10,Plot=FALSE) 
}
