% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialfun.r
\name{RS}
\alias{RS}
\title{\code{RS} rescaled adjusted range statistic}
\usage{
RS(x, k)
}
\arguments{
\item{x}{univariate time series}

\item{k}{length of the segments for which the statistic is computed. Starting with t=1, the segments do not overlap.}
}
\value{
(l,3)-matrix, 1. column: k, second column: starting time of segment, third column: value of RS statistic.
}
\description{
\code{RS} rescaled adjusted range statistic
}
\examples{
 
 data(TREMOR)
 R <- RS(TREMOR,10)  
}
