% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{SALES}
\alias{SALES}
\title{Monthly sales of a company}
\format{SALES is a univariate time series of length 77:
\describe{
\item{y}{monthly  sales of a company}
}}
\source{
Newton, H. J. (1988, ISBN: 978-0534091989): "TIMESLAB: A time series analysis laboraty"
}
\usage{
SALES
}
\description{
Monthly sales of a company
}
\examples{
data(SALES)
## maybe  tsp(SALES) ; plot(SALES) 
}
\keyword{datasets}
