% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{SCHAUINSLAND}
\alias{SCHAUINSLAND}
\title{CO2-Concentration obtained in Schauinsland, Germany}
\format{SCHAUINSLAND is a univariate time series of length 72:
\describe{
\item{SCHAUINSLAND}{CO2-Concentration obtained in Schauinsland}
}}
\source{
<http://cdiac.ornl.gov/trends/co2/uba/uba-sc.html>
}
\usage{
SCHAUINSLAND
}
\description{
CO2-Concentration obtained in Schauinsland, Germany
}
\examples{
data(SCHAUINSLAND)
## maybe  tsp(SCHAUINSLAND) ; plot(SCHAUINSLAND) 
}
\keyword{datasets}
