% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp.r
\name{smoothls}
\alias{smoothls}
\title{\code{smoothls} smoothes a time series by Whittaker graduation.  
The function depends on  the package Matrix.}
\usage{
smoothls(y, beta = 0)
}
\arguments{
\item{y}{the series, a vector or a time series}

\item{beta}{smoothing parameter  >=0 (the larger beta is, the smoother will g be)}
}
\value{
g     vector,  smooth component
}
\description{
\code{smoothls} smoothes a time series by Whittaker graduation.  
The function depends on  the package Matrix.
}
\examples{
 
data(GDP)
g <- smoothls(GDP,12)
\donttest{
 plot(GDP)   
 t <- seq(from = tsp(GDP)[1], to = tsp(GDP)[2],by=1/tsp(GDP)[3]) ; lines(t,g,col="red") } 
}
