% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequdom.r
\name{specplot}
\alias{specplot}
\title{\code{specplot}    plot of spectral estimate}
\usage{
specplot(s, Log = FALSE)
}
\arguments{
\item{s}{(n,2) or (n,4) matrix, output of specest}

\item{Log, }{logical, if TRUE, the logs of the spectral estimates are shown}
}
\description{
\code{specplot}    plot of spectral estimate
}
\examples{
\donttest{
data(WHORMONE)
est <- specest(WHORMONE,50,0.05,win = c("perwinba","perwinpa"),conf=0,type="cov") 
specplot(est,Log=FALSE) }
}
