% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{SPRUCE}
\alias{SPRUCE}
\title{Annual logging of spruce wood.}
\format{
SPRUCE is a univariate time series of length 42: 
\describe{
\item{SPRUCE}{Annual logging of spruce wood}
}
}
\usage{
SPRUCE
}
\description{
Annual logging of spruce wood.
}
\examples{
data(SPRUCE)
## maybe  tsp(SPRUCE) ; plot(SPRUCE) 
}
\keyword{datasets}
