% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smaniplot.r
\name{statcheck}
\alias{statcheck}
\title{\code{statcheck} determines the means, standard deviations and acf's of segmets of a time series 
                       and plots the acf's for the segments.}
\usage{
statcheck(y, d)
}
\arguments{
\item{y}{the series, a vector or a time series}

\item{d}{scalar, number of segments}
}
\value{
out list with components:
\item{ms}{matrix with means and standard deviations of the segments  }
\item{ac}{matrix with acf's, the first column: acf of the series, the others: acf's  of the segments  }
}
\description{
\code{statcheck} determines the means, standard deviations and acf's of segmets of a time series 
                       and plots the acf's for the segments.
}
\examples{
\donttest{
data(COFFEE)  
out <- statcheck(COFFEE,4)   }
}
