% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.r
\name{subsets}
\alias{subsets}
\title{\code{subsets} determines all subsets of  a set of n elements (labelled by  1,2,...,n ).}
\usage{
subsets(n)
}
\arguments{
\item{n}{scalar, integer >= 1}
}
\value{
mat (2^n,n)-matrix, each row gives the membership indicators of the elements 1,2,...,n
}
\description{
\code{subsets} determines all subsets of  a set of n elements (labelled by  1,2,...,n ).
}
\examples{
out <- subsets(4)

}
