% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smaniplot.r
\name{symplot}
\alias{symplot}
\title{\code{symplot} produces a symmetry plot}
\usage{
symplot(y)
}
\arguments{
\item{y}{the series, a vector or a time series}
}
\description{
\code{symplot} produces a symmetry plot
}
\examples{
\donttest{
data(LYNX)
symplot(LYNX)   }
}
