% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequdom.r
\name{taper}
\alias{taper}
\title{\code{taper} taper modification of a time series}
\usage{
taper(y, part)
}
\arguments{
\item{y}{the time series}

\item{part}{scalar, 0 <= part <= 0.5, part of modification (at each end of y)}
}
\value{
tp   tapered time series
}
\description{
\code{taper} taper modification of a time series
}
\examples{
data(WHORMONE)
out <-taper(WHORMONE,0.3)
\donttest{ 
plot(WHORMON) 
lines(out,col="red") }
}
