% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{TREERING}
\alias{TREERING}
\title{Mean thickness of annual tree rings}
\format{TREERING is a multivariate time series of length 66  with 3 variables:
\describe{
\item{THICK}{mean thickness of annual tree rings}                                 
\item{TEMP}{mean temperature of the year}             
\item{RAIN}{amount of rain of the year}
}}
\source{
<https://ltrr.arizona.edu/>
}
\usage{
TREERING
}
\description{
Mean thickness of annual tree rings
}
\examples{
data(TREERING)
## maybe  tsp(TREERING) ; plot(TREERING) 
}
\keyword{datasets}
