% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{TREMOR}
\alias{TREMOR}
\title{Measurements of physiological tremor}
\format{TREMOR is a univariate time series of length 400.
\describe{
\item{TREMOR}{Tremor}
}}
\usage{
TREMOR
}
\description{
Measurements of physiological tremor
}
\examples{
data(TREMOR)
## maybe  tsp(TREMOR) ; plot(TREMOR) 
}
\keyword{datasets}
