% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.r
\name{tsmat}
\alias{tsmat}
\title{\code{tsmat} constructs a (n-p+1,p) matrix from a time series 
where the first column is the shortened series y[p],...,y[n], the second is y[p-1],...,y[n-1],  etc.}
\usage{
tsmat(y, p)
}
\arguments{
\item{y}{the series, a vector or a time series of length n}

\item{p}{desired number of columns}
}
\value{
mat (n-p+1,p) matrix
}
\description{
\code{tsmat} constructs a (n-p+1,p) matrix from a time series 
where the first column is the shortened series y[p],...,y[n], the second is y[p-1],...,y[n-1],  etc.
}
\examples{
out <- tsmat(c(1:20),4)

}
