% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{USAPOP}
\alias{USAPOP}
\title{Population of USA}
\format{USAPOP is a univariate time series  of length 39; start 1630, frequency = 0.1 
\describe{
\item{USAPOP}{Population of USA}
}}
\source{
<https://www.worldometers.info/world-population/us-population/>
}
\usage{
USAPOP
}
\description{
Population of USA
}
\examples{
data(USAPOP)
## maybe  tsp(USAPOP) ; plot(USAPOP) 
}
\keyword{datasets}
