% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smaniplot.r
\name{vartable}
\alias{vartable}
\title{\code{vartable} determines table of variate differences}
\usage{
vartable(y, season)
}
\arguments{
\item{y}{the series, a vector or a time series}

\item{season}{scalar, period of seasonal component}
}
\value{
d      matrix with ratios of variances for differend numbers of simple and seasonal differencing
}
\description{
\code{vartable} determines table of variate differences
}
\examples{
data(GDP)
out <- vartable(GDP,4) 

}
