% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.r
\docType{data}
\name{WHORMONE}
\alias{WHORMONE}
\title{Concentration of growth hormone of a bull}
\format{WHORMONE is a univariate time series of length 97:
\describe{
\item{WHORMONE}{Concentration of growth hormone of a bull}
}}
\source{
Newton, H. J. (1988, ISBN: 978-0534091989): "TIMESLAB: A time series analysis laboraty"
}
\usage{
WHORMONE
}
\description{
Concentration of growth hormone of a bull
}
\examples{
data(WHORMONE)
## maybe  tsp(WHORMONE) ; plot(WHORMONE) 
}
\keyword{datasets}
