% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequdom.r
\name{wntest}
\alias{wntest}
\title{\code{wntest}   graphical test for white noise for a time series  or a series of regression residuals}
\usage{
wntest(e, a, k = 0)
}
\arguments{
\item{e}{vector, the time series (k = 0) or residuals (k > 0)}

\item{a}{scalar, level of significance}

\item{k}{scalar >= 0, number of regressors used to compute e as residuals}
}
\value{
tp vector,  value of test statistic and  p-value
}
\description{
\code{wntest}   graphical test for white noise for a time series  or a series of regression residuals
}
\examples{
\donttest{
data(WHORMONE)
out <- wntest(WHORMONE,0.05,0) }
}
